pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "birka-rag"
version = "0.1.0"
description = "RAG-система для работы с нормативными документами (эмбеддинги, Chroma, LLM)"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = {text = "Proprietary"}
authors = [
    {name = "Birka Team"}
]
keywords = ["rag", "llm", "embeddings", "chroma", "gigachat"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "pypdf>=3.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-huggingface>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-ollama>=0.2.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
openai = [
    "langchain-openai>=0.2.0",
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
birka-rag = "birka_rag.cli.chat:main"
birka-rag-index = "birka_rag.cli.index:main"

[project.urls]
Homepage = "https://github.com/birka/birka-rag"
Repository = "https://github.com/birka/birka-rag"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
birka_rag = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
